pyproject.toml
[tool.poetry]
name = "template-table-extractor"
version = "0.1.0"
description = "A tool and library to extract tables from pdf document based on templates."
authors = ["Sumit Arora <singhsays@gmail.com>"]
readme = "README.md"
packages = [{ include = "extractor" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.6"
loguru = "^0.7.0"
dacite = "^1.8.1"
pandas = "^2.0.3"
camelot-py = "^0.11.0"
opencv-python-headless = "^4.8.0.74"

[tool.poetry.scripts]
tablextractor = 'extractor:cli'

[tool.poetry.group.extras.dependencies]
pytest = "^7.4.0"
ruff = "^0.0.280"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    #"D", # pydocstyle
    "E",    # pycodestyle
    "W",    # pycodestyle
    "F",    # pyflakes
    "I001", # isort
    "UP",   # pyupgrade
]

[tool.pytest.ini_options]
minversion = "7.4"
norecursedirs = [".git", ".venv"]
testpaths = ["tests"]
